token expiry for employee fixes

+++ src/utils/axiosInstance.ts
@@ -39,7 +39,7 @@ API.interceptors.request.use(
 API.interceptors.response.use(
   (response) => response,
   (error) => {
-    if (error?.response?.status === 401) {
+    if (error?.response?.status === 401 && error?.response?.status === 403 ) {
       localStorage.clear();           // remove all tokens
       window.location.href = "/login";
     }
